add method_goto_installer
This is like the original method_installer, but instead of creating
a new sub that calls the original, it creates a new sub that jumps
into the original.  This leaves the normal caller() chain intact,
which is useful when the wrapped subroutine cares about its caller.

This was originally added so that you could use method_installer (well,
method_goto_installer) as the installer for Sub::Exporter.